Add Assertions property to ResultNode and use it in existing code
